refactor: rename hack to what it is
