+++ projects/client/src/lib/components/buttons/ActionButton.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import { useActiveLink } from "$lib/stores/useActiveLink";
   import { disableTransitionOn } from "$lib/utils/actions/disableTransitionOn";
+  import { mobileAppleDeviceTriggerHack } from "$lib/utils/actions/mobileAppleDeviceTriggerHack";
   import { triggerWithKeyboard } from "$lib/utils/actions/triggerWithKeyboard";
-  import { triggerWithTouch } from "$lib/utils/actions/triggerWithTouch";
   import type { TraktActionButtonProps } from "./TraktActionButtonProps";
 
   type TraktActionButtonAnchorProps = HTMLAnchorProps & TraktActionButtonProps;
@@ -25,7 +25,7 @@
   <a
     use:disableTransitionOn={"touch"}
     use:triggerWithKeyboard
-    use:triggerWithTouch
+    use:mobileAppleDeviceTriggerHack
     data-sveltekit-keepfocus
     data-sveltekit-noscroll={noscroll}
     class="trakt-action-button trakt-button-link"
