+++ projects/client/src/lib/components/dropdown/DropdownItem.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
+  import { mobileAppleDeviceTriggerHack } from "$lib/utils/actions/mobileAppleDeviceTriggerHack";
   import { triggerWithKeyboard } from "$lib/utils/actions/triggerWithKeyboard";
-  import { triggerWithTouch } from "$lib/utils/actions/triggerWithTouch";
   import type { Snippet } from "svelte";
   import Link from "../link/Link.svelte";
 
@@ -36,7 +36,7 @@
 <!-- svelte-ignore a11y_no_noninteractive_tabindex -->
 <li
   use:triggerWithKeyboard
-  use:triggerWithTouch
+  use:mobileAppleDeviceTriggerHack
   tabindex={tabIndex}
   data-color={color}
   {...props}
