+++ projects/client/src/lib/utils/actions/mobileAppleDeviceTriggerHack.ts
